Refactoring round 1
* Moved folders into normal folder locations.
* Refactored the actual class names to collapse the hierarchy, making
them easier to work with in JS.
* Replaced ApplicationFaade with Application, which instantiates a
framework Facade instance, registers StartupCommand.
* Changed the way the startup works, and the order the class files are
referenced in index.html
* Added a ChangeObject VO instead of passing an arbitrary object.
* Tidied the code.

There is more to go, the main thing we need to do is get the PageProxy
not doing anything with the View. In MVC, the Model should NEVER know
anything about the View. And that PageProxy should define all the notes
it sends, and not reference the AppConstants, which are really for the
View and Controller region - always remember, the Model should not know
anything about the app, including the constants it uses